Uses https for pypln
The configuration file is pointing to the place letsencrypt creates the
certificates. We chose to do this because letsencrypt is free, easy to use and
supported by all major browsers.

To create a new certificate just follow the documentation on
https://letsencrypt.readthedocs.org/en/latest/using.html . The domain for the
certificate will be assumed to be the one use to make the deploy (the same way
we assume that for the django sites initial data).

This commit enables HSTS and adds some specific security configurations, mostly
based on https://github.com/ioerror/duraconf/ .