Redirects from http to https using 307 instead of 301
Even though 307 is a temporary redirect (and this should be a permanent
redirect), returning 301 makes most clients change the HTTP method after the
redirect. This causes the document creation API to fail when users use the
`http` protocol in urls. Returning 308 would be a more suitable solution, but
support is apparently not widespread, so for now we will return 307 to
guarantee that the HTTP method is honored after redirects.

This fixes NAMD/pypln.web#139